Ensure rating icons do not overlap post reply ones
Changed height of the thumbrating div to match the size of its
contents. This div contains a rating-label div to show the number
of votes per comment and two <a> elements for the thumb icons.
Before, the height was just that of the rating-label, and did not
include the height of the thumb icons. Now, the height is correct
so that the icons do not overlap. Note that a height of auto could
also fix the same issue.