Hide ul when children li elements are hidden
li elements are hidded dynamically based on what is input into the
search box (lfr-menu-list-search-container). Previously there was
a problem when all the li elements were hidden (no matches for the
search) where the container was not hidden and appeared as a small
white box. This change will hide or display the ul as needed after
the search occurs.