LPS-45658 Dialog Iframe has too much body padding
For widths under 767px, body padding is added to the dialog iframe element
unecessarily. This rule is from a scss file from alloy bootstrap, and is
included in aui.css with the line '@import "aui/responsive"'. (See:
https://github.com/liferay/alloy-bootstrap/blob/master/lib/_responsive-767px-max.scss)
A rule is added to this file to override that padding for the dialog
iframe windows in the _styled theme and then its children